chore: moved utils to plugins/utils for reusability
