+++ plugins/utils/jsxImagesToChildren.js
@@ -4,8 +4,8 @@ const {
   parseImport,
   findAttribute,
   removeAttribute,
-} = require('../../../codemods/utils/mdxast');
-const { mdxBlockElement } = require('../../../codemods/utils/mdxast-builder');
+} = require('../../codemods/utils/mdxast');
+const { mdxBlockElement } = require('../../codemods/utils/mdxast-builder');
 const toMDAST = require('remark-parse');
 const remarkMdx = require('remark-mdx');
 const remarkMdxjs = require('remark-mdxjs');
